Improve data rendering
