+++ index.html
@@ -81,6 +81,7 @@
 			<model-selector />
 		</div>
 		<hr />
+		<div id="render_menu"></div>
 		<div id="plot-area">
 			<svg><g class="flip" style="transform: scale(1, -1) translate(0, -100%)"></g></svg>
 		</div>
