+++ js/data/base.js
@@ -58,6 +58,14 @@ export class BaseData {
 		return this.x.length
 	}
 
+	get columnNames() {
+		const names = []
+		for (let i = 0; i < this.dimension; i++) {
+			names[i] = `${i}`
+		}
+		return names
+	}
+
 	get x() {
 		return this._x
 	}
@@ -101,10 +109,6 @@ export class BaseData {
 	get isSeries() {
 		const task = this.setting.vue.mlTask
 		return ['TP', 'SM', 'CP'].indexOf(task) >= 0
-	}
-
-	get selectedColumnIndex() {
-		return this.isSeries ? [Math.min(1, this.dimension - 1)] : this.dimension === 1 ? [0] : [0, 1]
 	}
 
 	get params() {
@@ -202,93 +206,6 @@ export class MultiDimensionalData extends BaseData {
 
 		this._categorical_output = false
 		this._output_category_names = null
-
-		this._select = null
-	}
-
-	get selectedColumnIndex() {
-		return this._select?.() ?? super.selectedColumnIndex
-	}
-
-	_make_selector(names) {
-		let e = this.setting.data.configElement.select('div.column-selector')
-		if (e.size() === 0) {
-			e = this.setting.data.configElement.append('div').classed('column-selector', true)
-		} else {
-			e.selectAll('*').remove()
-		}
-		if (this.dimension <= 2) {
-			this._select = null
-		} else if (this.dimension <= 4) {
-			const elm = e.append('table').style('border-collapse', 'collapse')
-			let row = elm.append('tr').style('text-align', 'center')
-			row.append('td')
-			row.append('td').text('>')
-			row.append('td').text('V').style('transform', 'rotate(180deg')
-			const ck1 = []
-			const ck2 = []
-			for (let i = 0; i < this.dimension; i++) {
-				row = elm.append('tr')
-				elm.append('td').text(names[i]).style('text-align', 'right')
-				const d1 = elm
-					.append('td')
-					.append('input')
-					.attr('type', 'radio')
-					.attr('name', 'data-d1')
-					.on('change', () => this._manager.platform.render())
-				ck1.push(d1)
-				const d2 = elm
-					.append('td')
-					.append('input')
-					.attr('type', 'radio')
-					.attr('name', 'data-d2')
-					.on('change', () => this._manager.platform.render())
-				ck2.push(d2)
-			}
-			ck1[0].property('checked', true)
-			ck2[1].property('checked', true)
-			this._select = () => {
-				const k = []
-				for (let i = 0; i < this.dimension; i++) {
-					if (ck1[i].property('checked')) {
-						k[0] = i
-					}
-					if (ck2[i].property('checked')) {
-						k[1] = i
-					}
-				}
-				return k
-			}
-		} else {
-			names = names.map(v => '' + v)
-			e.append('span').text('>')
-			const slct1 = e.append('select').on('change', () => this._manager.platform.render())
-			slct1
-				.selectAll('option')
-				.data(names)
-				.enter()
-				.append('option')
-				.attr('value', d => d)
-				.text(d => d)
-			slct1.property('value', names[0])
-			e.append('span').text('V').style('transform', 'rotate(180deg').style('display', 'inline-block')
-			const slct2 = e.append('select').on('change', () => this._manager.platform.render())
-			slct2
-				.selectAll('option')
-				.data(names)
-				.enter()
-				.append('option')
-				.attr('value', d => d)
-				.text(d => d)
-			slct2.property('value', names[1])
-			this._select = () => [names.indexOf(slct1.property('value')), names.indexOf(slct2.property('value'))]
-		}
-		this._manager.platform.render()
-	}
-
-	terminate() {
-		this.setting.data.configElement.select('div.column-selector').remove()
-		super.terminate()
 	}
 }
