+++ js/platform/base.js
@@ -1,302 +1,5 @@
 import FittingMode from '../fitting.js'
-
-const scale = function (v, smin, smax, dmin, dmax) {
-	if (!isFinite(smin) || !isFinite(smax) || smin === smax) {
-		return (dmax + dmin) / 2
-	}
-	return ((v - smin) / (smax - smin)) * (dmax - dmin) + dmin
-}
-
-class DataRenderer {
-	constructor(manager) {
-		this._manager = manager
-		this._r = this.setting.svg.select('g.points g.datas')
-		if (this._r.size() === 0) {
-			const pointDatas = this.setting.svg.append('g').classed('points', true)
-			this._r = pointDatas.append('g').classed('datas', true)
-		}
-
-		this._p = []
-		this._pad = 10
-		this._clip_pad = -Infinity
-		this._pred_count = 0
-
-		this._observe_target = null
-		this._observer = new MutationObserver(mutations => {
-			if (this._observe_target) {
-				this._p.forEach(
-					(p, i) =>
-						(p.title = this.data._categorical_output
-							? this.data._output_category_names[this.data.y[i] - 1]
-							: this.data.y[i])
-				)
-			}
-		})
-		this._observer.observe(this.setting.svg.node(), {
-			childList: true,
-		})
-
-		this._will_render = false
-	}
-
-	get padding() {
-		if (!Array.isArray(this._pad)) {
-			return [this._pad, this._pad]
-		}
-		return this._pad
-	}
-
-	set padding(pad) {
-		this._pad = pad
-		this.render()
-	}
-
-	set clipPadding(pad) {
-		this._clip_pad = pad
-		this.render()
-	}
-
-	get _series() {
-		return this.data?.isSeries
-	}
-
-	get setting() {
-		return this._manager.setting
-	}
-
-	get width() {
-		return this._manager.platform.width
-	}
-
-	get height() {
-		return this._manager.platform.height
-	}
-
-	get points() {
-		return this._p
-	}
-
-	get data() {
-		return this._manager.datas
-	}
-
-	_clip(x) {
-		if (this._clip_pad === -Infinity) {
-			return x
-		}
-		const limit = [this.width, this.height]
-		for (let i = 0; i < x.length; i++) {
-			if (x[i] < this._clip_pad) {
-				x[i] = this._clip_pad
-			} else if (limit[i] - this._clip_pad < x[i]) {
-				x[i] = limit[i] - this._clip_pad
-			}
-		}
-		return x
-	}
-
-	render() {
-		if (!this._will_render) {
-			this._will_render = true
-			Promise.resolve().then(() => {
-				if (this._will_render) {
-					this._will_render = false
-					this._render()
-				}
-			})
-		}
-	}
-
-	toPoint(value) {
-		const k = this.data.selectedColumnIndex
-		const domain = this._series ? this.data.series.domain : this.data.domain
-		const range = [this.width, this.height]
-		const [ymin, ymax] = this.data.range
-		const d = k.map(
-			(t, s) => scale(value[t], domain[t][0], domain[t][1], 0, range[s] - this.padding[s] * 2) + this.padding[s]
-		)
-		if (this._series) {
-			if (Array.isArray(value[0])) {
-				const r = this.data.domain[0]
-				d[1] = scale(value[1], ymin, ymax, 0, range[1] - this.padding[1] * 2) + this.padding[1]
-				d[0] = scale(value[0][0], r[0], r[1], 0, range[0] - this.padding[0] * 2) + this.padding[0]
-			} else {
-				const k0 = Math.min(k[0], value[1].length - 1)
-				d[1] =
-					scale(value[1][k0], domain[k[0]][0], domain[k[0]][1], 0, range[1] - this.padding[1] * 2) +
-					this.padding[1]
-				d[0] =
-					scale(value[0], 0, this.data.length + this._pred_count, 0, range[0] - this.padding[0] * 2) +
-					this.padding[0]
-			}
-		}
-		if (d.length === 1 && value.length > 1) {
-			d[1] = scale(value[1], ymin, ymax, 0, range[1] - this.padding[1] * 2) + this.padding[1]
-		}
-		return d.map(v => (isNaN(v) ? 0 : v))
-	}
-
-	toValue(x) {
-		if (x && this._series) {
-			return [scale(x[0] - this.padding[0], 0, this.width - this.padding[0] * 2, 0, this.data.length)]
-		}
-		return []
-	}
-
-	_render() {
-		if (!this.data || this.data.dimension === 0) {
-			this._p.map(p => p.remove())
-			this._p.length = 0
-			return
-		}
-		const k = this.data.selectedColumnIndex
-		const n = this.data.length
-		const data = this._series ? this.data.series.values : this.data.x
-		const domain = this._series ? this.data.series.domain : this.data.domain
-		const range = [this.width, this.height]
-		const [ymin, ymax] = this.data.range
-		const radius = Math.max(1, Math.min(5, Math.floor(2000 / n)))
-		for (let i = 0; i < n; i++) {
-			const d = k.map(
-				(t, s) =>
-					scale(data[i][t], domain[t][0], domain[t][1], 0, range[s] - this.padding[s] * 2) + this.padding[s]
-			)
-			if (this._series) {
-				d[1] =
-					scale(data[i][k[0]], domain[k[0]][0], domain[k[0]][1], 0, range[1] - this.padding[1] * 2) +
-					this.padding[1]
-				d[0] = scale(i, 0, n + this._pred_count, 0, range[0] - this.padding[0] * 2) + this.padding[0]
-			}
-			if (d.length === 1) {
-				d[1] = scale(this.data.y[i], ymin, ymax, 0, range[1] - this.padding[1] * 2) + this.padding[1]
-			}
-
-			const dp = this._clip(d)
-			const cat = this.data.dimension === 1 ? 0 : this.data.y[i]
-			if (this._p[i]) {
-				const op = this._p[i].at
-				if (op[0] !== dp[0] || op[1] !== dp[1]) {
-					this._p[i].at = dp
-				}
-				if (this._p[i].category !== cat) {
-					this._p[i].category = cat
-				}
-			} else {
-				this._p[i] = new DataPoint(this._r, dp, cat)
-			}
-			this._p[i].title = this.data._categorical_output
-				? this.data._output_category_names[this.data.y[i] - 1]
-				: this.data.y[i]
-			this._p[i].radius = radius
-		}
-		for (let i = n; i < this._p.length; i++) {
-			this._p[i].remove()
-		}
-		this._p.length = n
-	}
-
-	predict(step) {
-		if (!Array.isArray(step)) {
-			step = [step, step]
-		}
-		const domain = this.data.domain
-		const range = [this.width, this.height]
-		const tiles = []
-		if (this.data.dimension <= 2) {
-			for (let i = 0; i < range[0] + step[0]; i += step[0]) {
-				const w = scale(i - this.padding[0], 0, range[0] - this.padding[0] * 2, domain[0][0], domain[0][1])
-				if (this.data.dimension === 1) {
-					tiles.push([w])
-				} else {
-					for (let j = 0; j < range[1] - step[1] / 100; j += step[1]) {
-						const h = scale(
-							j - this.padding[1],
-							0,
-							range[1] - this.padding[1] * 2,
-							domain[1][0],
-							domain[1][1]
-						)
-						tiles.push([w, h])
-					}
-				}
-			}
-		} else {
-			for (let i = 0; i < this.data.x.length; i++) {
-				tiles.push(this.data.x[i].concat())
-			}
-		}
-		const task = this.setting.vue.mlTask
-		const plot = (pred, r) => {
-			r.selectAll('*').remove()
-			let smooth = pred.some(v => !Number.isInteger(v))
-			if (this.data.dimension === 1) {
-				const p = []
-				if (task === 'IN' || (smooth && task !== 'DE')) {
-					const [ymin, ymax] = this.data.range
-					for (let i = 0; i < pred.length; i++) {
-						p.push([
-							scale(tiles[i], domain[0][0], domain[0][1], 0, range[0] - this.padding[0] * 2) +
-								this.padding[0],
-							scale(pred[i], ymin, ymax, 0, range[1] - this.padding[1] * 2) + this.padding[1],
-						])
-					}
-
-					const line = d3
-						.line()
-						.x(d => d[0])
-						.y(d => d[1])
-					r.append('path').attr('stroke', 'red').attr('fill-opacity', 0).attr('d', line(p))
-				} else {
-					p.push([], [])
-					for (let i = 0, w = 0; w < range[0] + step[0]; i++, w += step[0]) {
-						p[0][i] = pred[i]
-						p[1][i] = pred[i]
-					}
-
-					const t = r.append('g')
-					new DataHulls(t, p, [step[0], 1000], smooth)
-				}
-			} else if (this.data.dimension === 2) {
-				let c = 0
-				const p = []
-				for (let i = 0, w = 0; w < range[0] + step[0]; i++, w += step[0]) {
-					for (let j = 0, h = 0; h < range[1] - step[1] / 100; j++, h += step[1]) {
-						if (!p[j]) p[j] = []
-						p[j][i] = pred[c++]
-					}
-				}
-				if (!smooth && pred.length > 100) {
-					smooth |= new Set(pred).size > 100
-				}
-
-				const t = r.append('g')
-				new DataHulls(t, p, step, smooth || task === 'DE')
-			} else {
-				const t = r.append('g')
-				const name = pred.every(Number.isInteger)
-				for (let i = 0; i < pred.length; i++) {
-					const o = new DataCircle(t, this._p[i])
-					o.color = getCategoryColor(pred[i])
-					if (name && this.data._categorical_output) {
-						this._p[i].title = `true: ${this.data._output_category_names[this.data.y[i] - 1]}\npred: ${
-							this.data._output_category_names[pred[i] - 1]
-						}`
-					} else {
-						this._p[i].title = `true: ${this.data.y[i]}\npred: ${pred[i]}`
-					}
-				}
-				this._observe_target = r
-			}
-		}
-		return [tiles, plot]
-	}
-
-	terminate() {
-		this._p.forEach(p => p.remove())
-		this._observer.disconnect()
-		this._will_render = false
-	}
-}
+import DataRenderer from '../renderer/plot.js'
 
 export class BasePlatform {
 	constructor(task, manager) {
@@ -462,6 +165,7 @@ export class DefaultPlatform extends BasePlatform {
 			.attr('opacity', renderFront ? 1 : 0.5)
 		this.setting.footer.text('')
 		this.svg.select('g.centroids').remove()
+		this._renderer.init()
 		this.render()
 	}
